musicians/src/requirements.txt
django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ
psycopg2-binary
django-autoslug
django-embed-video
pillow
requests
